% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{autumn}
\alias{autumn}
\title{Picture of an autumn forest.}
\format{
An object of class \code{stars}.
}
\source{
private archive
}
\usage{
autumn
}
\description{
A picture of an autumn forest in north-east Bavaria ("Fichtelgebirge").
}
\keyword{datasets}
