% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{mapLayers}
\alias{mapLayers}
\title{Query a set of map layers from the Overpass API}
\usage{
mapLayers(bbox, layertags)
}
\arguments{
\item{bbox, }{an overpass_query object.}

\item{layertags, }{an object of class \code{mapLayerTags} listing all
required map layers and associated feature tags. Typically created by
\code{\link[=mapLayerTags]{mapLayerTags()}}.}
}
\value{
A list of \code{sf} objects.
}
\description{
Query a set of map layers from the Overpass API
}
\examples{
library(osmdata)
bbox = getbb("Forchheim")
layertags = mapLayerTags(
  maptype = NULL
  , layers = c("roadsMajor", "roadsMinor")
)
mapLayers(bbox, layertags = layertags)

}
