% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{mapLayerTags}
\alias{mapLayerTags}
\title{Create a feature list for a series of map layers.}
\usage{
mapLayerTags(maptype = "building", layers = NULL)
}
\arguments{
\item{maptype, }{a \code{character} indicating the default map type to be created.
Currently supports "building" and "street". If set to \code{NULL}, a set of
layers can be specified through \code{layers}.}

\item{layers, }{a \code{character} vector specifying the layers to be included in
the map output. Can be any combination of "buildings", "roadsMajor",
"roadsMinor", "roadsSmall", "rivers" and "railways". For each of these
layers a standard set of features is included.}
}
\value{
A nested \code{list} with layers and key-value-lists per layer.
}
\description{
Create a feature list for a series of map layers.
}
\details{
This function consolidates feature tags for pre-defined map layers into a
list.
}
\examples{
mapLayerTags(maptype = "building")
mapLayerTags(maptype = "street")
mapLayerTags(maptype = NULL, layers = c("buildings"))

}
