% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{mapOSM}
\alias{mapOSM}
\title{Create a ggplot map object}
\usage{
mapOSM(
  location,
  maptype = "building",
  design = "night",
  options = mapOptions(maptype, design = design),
  bg = "black",
  title = NULL,
  subitle = NULL
)
}
\arguments{
\item{layers, }{a list of \code{sf} objects, typically generated by \code{\link[=mapLayers]{mapLayers()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Create a ggplot map object
}
\examples{
library(osmdata)
bbox = getbb("Forchheim")
layertags = mapLayerTags(
  maptype = NULL
  , layers = c("roadsMajor", "roadsMinor")
)
layers = mapLayers(bbox, layertags = layertags)
mapOSM(layers, bg = "black", col = "white")

}
