% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mountains}
\alias{mountains}
\title{Picture of a mountains.}
\format{
An object of class \code{stars}.
}
\source{
private archive
}
\usage{
mountains
}
\description{
A picture of the "Großglockner", hightes mountain of Austria.
}
\keyword{datasets}
