% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painting.R
\name{painting}
\alias{painting}
\title{Create a painting from a picture using supercells}
\usage{
painting(x, ...)
}
\arguments{
\item{x, }{object of class \code{stars}, the picture to be processed.}

\item{...}{further arguments passed to \code{supercells()}}
}
\value{
Object of class \code{sf} containing the boundaries of the supercells and the
average hex color value.
}
\description{
Create a painting from a picture using supercells
}
\details{
This function takes an image and applies the concept of superpixels as
implemented in the \href{https://github.com/Nowosad/supercells}{supercells}
package to get a painting like representation.
}
\examples{
data(autumn)
ptg = painting(pic)
par(mfrow = c(1, 2))
plot(pic, rgb = 1:3)
plot(ptg)

}
