% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painting.R
\name{plot.sf_paint}
\alias{plot.sf_paint}
\title{Plot object of class \code{sf_paint}}
\usage{
\method{plot}{sf_paint}(x, col = x$col, ...)
}
\arguments{
\item{x, }{object of class \code{sf_paint}.}

\item{col, }{vector of color values. Defaults to the hexcolor column
attached when running \code{painting()}.}
}
\description{
Plot object of class \code{sf_paint}
}
