% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painting.R
\name{rgb2hex}
\alias{rgb2hex}
\title{Create hex color values from RGB}
\usage{
rgb2hex(x)
}
\arguments{
\item{x, }{\code{matrix} or \code{data.frame} of rgb color values.}
}
\value{
A \code{character} vector of hex color values
}
\description{
Create hex color values from RGB
}
