% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painting.R
\name{painting}
\alias{painting}
\title{Create a painting from a picture using supercells}
\usage{
painting(x, ...)
}
\arguments{
\item{x, }{object of class \code{stars}, the picture to be processed.}

\item{...}{further arguments passed to \code{supercells()}}
}
\value{
<<<<<<< HEAD
Object of class \code{sf} containing the boundaries of the supercells and the
=======
Object of class \code{sf} containig the boundaries of the supercells and the
>>>>>>> osmmaps
average hex color value.
}
\description{
Create a painting from a picture using supercells
}
\details{
This function takes an image and applies the concept of superpixels as
implemented in the \href{https://github.com/Nowosad/supercells}{supercells}
package to get a painting like representation.
}
\examples{
ifl = system.file("img/lake.jpg", package = "aRts")
lake = stars::read_stars(ifl)
lake_ptg = painting(lake)
par(mfrow = c(1, 2))
plot(lake, rgb = 1:3, main = "", reset = FALSE, useRaster = TRUE)
plot(lake_ptg)

}
